Refactor: Rename `docbuild.cli.config`
Rename it to `docbuild.cli.cmd_config` to make it consistent with
other CLI commands.
