+++ src/docbuild/cli/__init__.py
@@ -2,8 +2,10 @@
 
 * :class:`~docbuild.cli.cmd_build`: Build the documentation.
 * :class:`~docbuild.cli.cmd_c14n`: Canonicalize the documentation.
+* :class:`~docbuild.cli.cmd_config`: Configuration management for docbuild CLI commands.
 * :class:`~docbuild.cli.cmd_cli`: Main CLI entry point for docbuild.
-* :class:`~docbuild.cli.config`: Configuration management for docbuild CLI commands.
+* :class:`~docbuild.cli.cmd_metadata`: Metadata management for docbuild.
+* :class:`~docbuild.cli.cmd_validate`: Validate the documentation.
 
 Additional modules:
 * :class:`~docbuild.cli.context`: Context for the docbuild CLI commands.
